Now listing schools inside fragment.
Tossed ClassListView.java as it was way more work that it was worth to
implement.  MainActivity.java now uses a fragment to display the list of
schools.  SchoolListFragment.java is the fragment that displays the list of
schools.  school_list.xml is the TextView used in the lists.  School.java
overrides the toString() method to return the name of the school.