Major overhaul of network code.
Added in a class that overrides the string request class from volley.  Now the
NetworkRequest class fills in the missing parts for the network code.  It is
much more generic and easier to follow.  Also added checks for error status
codes, though still need to figure out how to pass that along to the fragments.